pyproject.toml
[tool.poetry]
name = "chartova"
version = "0.1.0"
description = ""
authors = ["Alexander <csmcgl@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.0,<3.13.0"
numpy = "^1.26.1"
pandas = "^2.1.2"
matplotlib = "^3.8.1"
celery = "^5.3.4"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
fake-useragent = "^1.3.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
